pyproject.toml
[project]
name = "mlflow-demos"
version = "0.1.0"
description = "MLflow demo agents for DevConnect"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mlflow==3.10.0",
    "openai>=1.0.0",
    "litellm>=1.0.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "ipython>=8.0.0",
    "ipykernel>=7.2.0",
]

[project.scripts]
mlflow-restaurant-research-bot = "devconnect.restaurant_research_bot.restaurant_research_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["devconnect"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]
